handler-py/pyproject.toml
[project]
name = "rocketnotes_handler"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12.9"
dependencies = [
    "boto3==1.38.8",
    "httpx>=0.28.1",
    "faiss-cpu==1.11.0",
    "langchain==0.3.24",
    "langchain-anthropic==0.3.12",
    "langchain-community==0.3.23",
    "langchain-core==0.3.56",
    "langchain-openai==0.3.14",
    "langchain-text-splitters==0.3.8",
    "langchain-together==0.3.0",
    "numpy>=2.2.5",
    "scikit-learn>=1.6.1",
]

[project.optional-dependencies]
dev = ["pytest==7.4.4", "aws-cdk-lib==2.190.0", "constructs>=10.0.0,<11.0.0"]

[tool.pytest.ini_options]
pythonpath = ["."]
